utils/pool/requirements.txt
tqdm~=4.64.0
requests~=2.27.1
PyYAML~=6.0
maxminddb~=2.2.0
setuptools~=60.2.0
emoji-country-flag~=1.3.0
